pundit exept destroy list need to test with not admin

+++ app/controllers/articles_controller.rb
@@ -3,6 +3,7 @@ class ArticlesController < ApplicationController
   def create
     @list = List.find(params[:list_id])
     @article = Article.new(article_params)
+    authorize @article
     if @article.save
       Show.create(article: @article, list: @list)
       redirect_to list_path(@list)
@@ -13,11 +14,13 @@ class ArticlesController < ApplicationController
 
   def edit
     @article = Article.find(params[:id])
+    authorize @article
     @list = List.find(params[:list_id])
   end
 
   def update
     @article = Article.find(params[:id])
+    authorize @article
     @list = List.find(params[:list_id])
     if @article.update(article_params)
       redirect_to list_path(@list)
@@ -27,9 +30,10 @@ class ArticlesController < ApplicationController
   end
 
   def destroy
-    article = Article.find(params[:id])
+    @article = Article.find(params[:id])
+    authorize @article
     list = List.find(params[:list_id])
-    show = article.shows.where(:list_id == params[:list_id])
+    show = @article.shows.where(:list_id == params[:list_id])
     show.first.destroy
     redirect_to list_path(list)
   end
